WIP refactor namespacing for federation libs, into Diaspora::Federated
