+++ lib/diaspora/federated.rb
@@ -1,0 +1,12 @@
+#   Copyright (c) 2010-2011, Diaspora Inc.  This file is
+#   licensed under the Affero General Public License version 3 or later.  See
+#   the COPYRIGHT file.
+
+module Diaspora
+  module Federated
+    require 'diaspora/federated/request'
+    require 'diaspora/federated/retraction'
+    require 'diaspora/federated/signed_retraction'
+    require 'diaspora/federated/relayable_retraction'
+  end
+end
