+++ spec/lib/diaspora/federated/request_spec.rb
@@ -3,16 +3,15 @@
 #   the COPYRIGHT file.
 
 require 'spec_helper'
-require Rails.root.join('lib', 'diaspora', 'federated', 'messages')
 
-describe Diaspora::Federated::Messages::Request do
+describe Diaspora::Federated::Request do
   before do
     @aspect = alice.aspects.first
   end
 
   describe 'validations' do
     before do
-      @request = Request.diaspora_initialize(:from => alice.person, :to => eve.person, :into => @aspect)
+      @request = described_class.diaspora_initialize(:from => alice.person, :to => eve.person, :into => @aspect)
     end
 
     it 'is valid' do
@@ -43,7 +42,7 @@ describe Diaspora::Federated::Messages::Request do
     end
 
     it 'is not to yourself' do
-      @request = Request.diaspora_initialize(:from => alice.person, :to => alice.person, :into => @aspect)
+      @request = described_class.diaspora_initialize(:from => alice.person, :to => alice.person, :into => @aspect)
       @request.should_not be_valid
     end
   end
@@ -52,7 +51,7 @@ describe Diaspora::Federated::Messages::Request do
     it 'returns request_accepted' do
       person = FactoryGirl.build:person
 
-      request = Request.diaspora_initialize(:from => alice.person, :to => eve.person, :into => @aspect)
+      request = described_class.diaspora_initialize(:from => alice.person, :to => eve.person, :into => @aspect)
       alice.contacts.create(:person_id => person.id)
 
       request.notification_type(alice, person).should == Notifications::StartedSharing
@@ -61,14 +60,14 @@ describe Diaspora::Federated::Messages::Request do
 
   describe '#subscribers' do
     it 'returns an array with to field on a request' do
-      request = Request.diaspora_initialize(:from => alice.person, :to => eve.person, :into => @aspect)
+      request = described_class.diaspora_initialize(:from => alice.person, :to => eve.person, :into => @aspect)
       request.subscribers(alice).should =~ [eve.person]
     end
   end
 
   describe '#receive' do
     it 'creates a contact' do
-      request = Request.diaspora_initialize(:from => alice.person, :to => eve.person, :into => @aspect)
+      request = described_class.diaspora_initialize(:from => alice.person, :to => eve.person, :into => @aspect)
       lambda{
         request.receive(eve, alice.person)
       }.should change{
@@ -80,7 +79,7 @@ describe Diaspora::Federated::Messages::Request do
       alice.share_with(eve.person, alice.aspects.first)
 
       lambda {
-        Request.diaspora_initialize(:from => eve.person, :to => alice.person,
+        described_class.diaspora_initialize(:from => eve.person, :to => alice.person,
                                     :into => eve.aspects.first).receive(alice, eve.person)
       }.should change {
         alice.contacts.find_by_person_id(eve.person.id).mutual?
@@ -89,7 +88,7 @@ describe Diaspora::Federated::Messages::Request do
     end
 
     it 'sets sharing' do
-      Request.diaspora_initialize(:from => eve.person, :to => alice.person,
+      described_class.diaspora_initialize(:from => eve.person, :to => alice.person,
                                   :into => eve.aspects.first).receive(alice, eve.person)
       alice.contact_for(eve.person).should be_sharing
     end
@@ -99,10 +98,10 @@ describe Diaspora::Federated::Messages::Request do
       alice.auto_follow_back_aspect = alice.aspects.first
       alice.save
       
-      Request.diaspora_initialize(:from => eve.person, :to => alice.person,
-                                  :into => eve.aspects.first).receive(alice, eve.person)
+      described_class.diaspora_initialize(:from => eve.person, :to => alice.person,
+                                          :into => eve.aspects.first).receive(alice, eve.person)
       
-      eve.contact_for(alice.person).should be_sharing
+      eve.contact_for( alice.person ).should be_sharing
     end
     
     it 'shares not back if auto_following is not enabled' do
@@ -110,7 +109,7 @@ describe Diaspora::Federated::Messages::Request do
       alice.auto_follow_back_aspect = alice.aspects.first
       alice.save
       
-      Request.diaspora_initialize(:from => eve.person, :to => alice.person,
+      described_class.diaspora_initialize(:from => eve.person, :to => alice.person,
                                   :into => eve.aspects.first).receive(alice, eve.person)
       
       eve.contact_for(alice.person).should be_nil
@@ -127,14 +126,14 @@ describe Diaspora::Federated::Messages::Request do
       
       alice.should_not_receive(:share_with)
       
-      Request.diaspora_initialize(:from => eve.person, :to => alice.person,
+      described_class.diaspora_initialize(:from => eve.person, :to => alice.person,
                                   :into => eve.aspects.first).receive(alice, eve.person)
     end
   end
 
   context 'xml' do
     before do
-      @request = Request.diaspora_initialize(:from => alice.person, :to => eve.person, :into => @aspect)
+      @request = described_class.diaspora_initialize(:from => alice.person, :to => eve.person, :into => @aspect)
       @xml = @request.to_xml.to_s
     end
 
@@ -149,7 +148,7 @@ describe Diaspora::Federated::Messages::Request do
 
     context 'marshalling' do
       it 'produces a request object' do
-        marshalled = Request.from_xml @xml
+        marshalled = described_class.from_xml @xml
 
         marshalled.sender.should == alice.person
         marshalled.recipient.should == eve.person
